+++ cpp/bindings/bindings.hpp
@@ -46,6 +46,8 @@ class PyPiecewisePolyPath {
   py::bytes serialize() const;
   // Deserialize  frm a bytes stream.
   void deserialize(const py::bytes&);
+  void constructHermite(const Vectors& positions, const Vectors& velocities,
+                        const std::vector<value_type> times);
 
   std::string __str__();
   std::string __repr__();