+++ cpp/bindings/toppra_int.cpp
@@ -30,6 +30,7 @@ PYBIND11_MODULE(toppra_int, m) {
       .def("eval", &PyPiecewisePolyPath::eval)
       .def("serialize", &PyPiecewisePolyPath::serialize)
       .def("deserialize", &PyPiecewisePolyPath::deserialize)
+      .def("constructHermite", &PyPiecewisePolyPath::constructHermite)
       .def("__call__", &PyPiecewisePolyPath::eval, py::arg("xs"), py::arg("order") = 0)
       .def("__str__", &PyPiecewisePolyPath::__str__)
       .def("__repr__", &PyPiecewisePolyPath::__repr__)