+++ HISTORY.md
@@ -5,7 +5,7 @@
 ### Added
 - [cpp] Implement serialization/deserialization for piecewise poly trajectory.
 - [cpp] Provide Python bindings for PiecewisePolyPath class.
-- [cpp] Implement a class for computing hermite spline.
+- [cpp] Construct piecewise poly as a hermite spline.
 
 ### Changed
 - [cpp] Minor improvement to PiecewisePolyPath.